+++ assets/js/components/Top/Notifications.vue
@@ -41,7 +41,7 @@
 								class="d-flex justify-content-end mt-3"
 								:title="fmtAbsoluteDate(msg.time)"
 							>
-								{{ fmtTimeAgo(msg.time - new Date()) }}
+								{{ fmtTimeAgo(msg.time.getTime() - new Date().getTime()) }}
 							</small>
 							<p class="d-flex align-items-baseline">
 								<shopicon-regular-exclamationtriangle
@@ -86,16 +86,24 @@
 	</div>
 </template>
 
-<script>
+<script lang="ts">
 import "@h2d2/shopicons/es/regular/exclamationtriangle";
 import formatter from "@/mixins/formatter";
+import { defineComponent, type PropType } from "vue";
+import type { Notification } from "./types";
+import type { Timeout } from "@/types/evcc";
 
-export default {
+export default defineComponent({
 	name: "Notifications",
 	mixins: [formatter],
 	props: {
-		notifications: Array,
-		loadpointTitles: Array,
+		notifications: { type: Array as PropType<Notification[]>, default: () => [] },
+		loadpointTitles: { type: Array as PropType<string[]>, default: () => [] },
+	},
+	data() {
+		return {
+			interval: null as Timeout,
+		};
 	},
 	computed: {
 		iconVisible() {
@@ -113,10 +121,12 @@ export default {
 		}, 10 * 1000);
 	},
 	unmounted() {
-		clearTimeout(this.interval);
+		if (this.interval) {
+			clearTimeout(this.interval);
+		}
 	},
 	methods: {
-		message({ message, lp }) {
+		message({ message, lp }: { message: string | string[]; lp?: number }) {
 			const lines = Array.isArray(message) ? message : [message];
 			if (lp) {
 				// add loadpoint title to first line
@@ -128,5 +138,5 @@ export default {
 			window.app?.clear();
 		},
 	},
-};
+});
 </script>
