+++ assets/js/components/Top/types.d.ts
@@ -1,0 +1,14 @@
+export interface Notification {
+	message: string;
+	time: Date;
+	level: string;
+	lp: number;
+	count: number;
+}
+
+export interface Provider {
+	title: string;
+	loggedIn: boolean;
+	loginPath: string;
+	logoutPath: string;
+}
