+++ assets/js/types/evcc.ts
@@ -7,6 +7,12 @@ declare global {
 	}
 }
 
+export interface Auth {
+	vehicles: VehicleLogins;
+}
+
+export type VehicleLogins = Record<string, { authenticated: boolean; uri: string }>;
+
 export interface FatalError {
 	error: any;
 	class?: any;
@@ -19,6 +25,7 @@ export interface State {
 	forecast?: Forecast;
 	currency?: CURRENCY;
 	fatal?: FatalError;
+	auth?: Auth;
 }
 
 export interface LoadpointCompact {
